- TextEdit: Minor cleanup and font connection fix.

+++ Applications/TextEdit/MultiplePageView.m
@@ -166,7 +166,7 @@
     return YES;     
 }
 
-- (NSRect)rectForPage:(int)page {
+- (NSRect)rectForPage:(NSInteger)page {
     return [self documentRectForPageNumber:page-1];  /* Our page numbers start from 0; the kit's from 1 */
 }
